--- src/test/java/com/donut/prokindonutsweb/controller/member/WarehouseControllerTest.java
@@ -1,4 +1,0 @@
-package com.donut.prokindonutsweb.controller.member;
-
-public class WarehouseControllerTest {
-}
