+++ src/test/java/com/donut/prokindonutsweb/controller/warehouse/WarehouseControllerTest.java
@@ -1,0 +1,4 @@
+package com.donut.prokindonutsweb.controller.warehouse;
+
+public class WarehouseControllerTest {
+}
